Forbid using global selection when editing transform masks
Using global selections in the transform tool when editing
transformation masks was an "accidental feature". Transform masks
were not designed to handle it properly. Using selections in
transformation  masks is a good idea, but it needs time for
implementing it. So right now we just disable using selections in
transform masks and postpone it till later :)

CCBUG:344196