BUG:184746 merging between two flake layers should result a flake layer
Added virtual KisLayer::createMergedLayer() and implemented it
for KisShapeLayer and KisGroupLayer. KisImage::mergeDown now
uses this method which now can be overridden by descendants
to specialize merge operations.

This fix seems to work, but I'm not sure if I've missed a corner
case or a use-case. Also, it might be useful to have a KoShape::clone(),
as the current code uses odf pasting to clone shapes, as that's
what the KisShapeLayer copy constructor did.

ccmail:kimageshop@kde.org