Adding tangent normal map blending mode to krita.
This patch adds a tangent normal map blending mode to Krita. It works by
taking the base value and substracting that from the destination before adding
it onto the source. It's still experimental in that it needs testing.

It's to be found under misc.

I had a huge fight with git to get my working branch synced to master,
and I am pretty sure it's clean now, but if I messed up the branch, feel free
to revert this commit.