Add experimental save/load support for animations.
The file format used is a simple extension of the .kra format, chosen for
ease of implementation. It is not intended as a final design but rather
to allow easier testing of the animation features until the actual format 
is decided upon.

A warning about this is included to inform the user when saving files.