Port filters/libmsooxml to Qt5/KF5.
This is the actual port, but it's not enabled yet in
CalligraProducts.cmake. The reason for that is that libs/main is not yet
ported, and we need to link against that. As far as I understand, once that is
done, the only thing needed will be to move the LIBMSOOXML product from
UNPORTED to PORTED.