The footnotes and endnotes were not displayed using the correct styles. We had just hard coded them to be shown using superscript, but they should be done using the styles specified by the notes configuration.
The same applies to the paragraph style of the note itself - but since this is just a default we didn't see any problem
when loading files, only when adding our own footnotes.

Now netiher of the 3 styles specified by the notes configuraion needs to be present in files, and indeed are not present in
our templates, so what LO does is add some styles. We do the same with these 6 styles

Footnote : based on default paragraph
Footnote Symbol : Applied to the id in the footnote itself - just a black text style
Footnote anchor : Applied to the id in the main text area - superscript
Endnote : based on default paragraph
Endnote Symbol : Applied to the id in the endnote itself - just a black text style
Endnote anchor : Applied to the id in the main text area - superscript

BUG:306141
REVIEW: 123195

This still doesn't work really well as the configuration stylenames are %encoded but the names in the stylemanager are
displaynames. That requires another fixup.