Fix finding of the cached thumbnails.
The actual image in the case might be smaller then the image size requested.
When adding the image to the change it might be have reduced in size to have
no empty space around it. Therefore save the expected size in the cache so it
gets found when it is smaller due to empty space around it.