Fix compilation on Windows
1) Unused but exported class definition makes MSVC cry
2) 'using' keyword makes the compiler get crazy. Don't use it and
   refactor the code in the future.

There is a really frightening warning as well, which is not yet fixed.
Though I guess it means copy-ctor and operator=() which are not used but
still generated. It must not access the destructor of the shared pointer
in the header since the destructor of the class is defined in the .cpp.

 83>c:\dev\desktop64\i\include\QtCore/qsharedpointer_impl.h(342): warning C4150: deletion of pointer to incomplete type 'KisPSDLayerStyle'; no destructor called [C:\dev\desktop64\p\vs11\krita\ui\kritaui.vcxproj]

Full log:
https://paste.kde.org/puwwwo8pe