Improve KoFindToolbar uasability
The KoFindToolbar is used in words to search the documents. The current behaviour
is a bit stange. When entering a string to search it waits until there is no
keypress for 1 sec before it starts to search. This also happens when you press
enter. On pressing enter it even goes to the last found string that was seached
last.

The patch changes the behaviour to start the search directly after pressing enter
without waiting 1 second. If the seach string was not changed enter still goes to
the next found match.

REVIEW: 123660