BUG:345754 120 and 300 degrees are different angles but also parallel.
As indicated by the commit message the issue was that in 360 degrees there's always
two values which can be parallel. Fixed with an fmod.

I can now not reproduce this bug any more. The downside of course is that
perspective squares with parallel sides don't do preview lines for that side, but we
can look at that in the future.