Fix preset docker flickering and slowdown
When vertically resizing the preset docker and the vertical
scrollbar appears, the preset tiles are resized to fit the
viewport with their width, but this also reduces their height
not needing the scrollbar anymore.
The scrollbar disappearing lets the viewport expand and so the tiles,
in height, making the scrollbar needed again and this loops forever
generating a considerable slowdown.
The fix consists in always showing the vertical scrollbar so that no
resizing happens due to them.

To avoid unnecessary work and event spam we check in
KoResourceItemChooser::setSynced function if the value passed
corresponds to its current status already.

Thanks to mmolch for the initial patch.