Adding a Fish Eye Vanishing Point assistant.
This adds the ability to draw perspective that look lens-deformed.
Like the vanishing point assistant, this assistant is per a set of parallel
lines in a 3d space. So to use it effectively, use two, where the second is at
a 90 degree angle of the first, and add a vanishing point to the center of
both. Or combine one with a parallel ruler and a vanishing point, or even
one with two vanishing points. The possibilities are quite large.

This assistant will not just give feedback/snapping between the vanishing
points, but also give feedback to the relative left and right of the assistant.
This is so you can use it in edge-cases like panoramas with relative ease.

Confirmed to work and to save and load, have fun!

CCMAIL:kimageshop@kde.org