Fix Adjustment layers composition
This was a very old bug related the way how selection is applied
to adjustment layers. Now selection defines not "how the filtered
data is blended into the image" but "how the filter effect is applied
to the composed image". In the result, erasing the selection now will
*not* make your image fully transparent, but just removes the filter
effect from that area.

BUG:324505,294122
CCBUG:349078
CC:kimageshop@kde.org