Make painting below Adjustment Layers up to 10 times faster! :)
We used to spend a lot of time on creation and update of progress
reporters. Now we use a specially factored class KisBusyProgressIndicator
that just shows a progress bar for 200ms and hides it if no further updates
happened. Usage of atomic integers allows us avoid expensive synchronisations
between threads.

So now painting below Adjustment layers is significantly faster!

CC:kimageshop@kde.org