BUG:348997 Don't crash when saving after saving failed
If autosave failed, the saver object wouldn't be deleted and we'd
show a messagebox. If you don't notice, after the next interval
passes, we'd autosave again and assert because the saver object
is not zero.

The same issue could conceivably happen when loading, which also
checked for errors and returned without deleting the loader object.

This patch always deletes the saver and loader object _and_ deletes
it if it still exists for whatever reason...