Fix selection bugs in the preset docker
Now is possible to select again a brush that is the only one in the
selected tag category.

When switching tag categories if the selected preset is present in the new
category, it will be selected, otherwise no preset will be, visually, selected,
but the active one won't be changed, so the user can still paint.

Added several checks to avoid unneeded executions and event emitting.
Especially now it shouldn't be possible to break the previous preset
selection/shortcut anymore.

BUG: 348120
BUG: 348444
Differential Revision: https://phabricator.kde.org/D124