Fix layer selection clearing when the model resets
Created slotModelReset slot function in KisLayerBox to substitute the modelReset signal
connection to updateUI so that is possible to restore the selection after the model resets.

Added some checks to avoid model reset when not needed.

Added hasDummiesFacade function to KisNodeModel so that it is possible to check when its dummiesFacade
is null due to code that temporarily unset it, to avoid using its functions and crash.

Added even more checks to avoid duplicated work when setting something that is
already set (setting the same canvas twice).

CCBUG: 349538