Restore loading of defaultstyles.xml
Summary:
Fetching the resource dirs for "styles" from the main componentdata failed, as the
resource dirs were registered by the factories with componentdata instances which
never become main components, as they are only created after the KApplication
instance, which itself already creates and registers a main componentdata .

Also did the old approach only work with one type of document. Which conflicts
at least with Gemini, which wants to support among others text documents
and presentations in the very same application.

Reviewers: yue, boemann

Reviewed By: yue, boemann

Subscribers: boemann, yue, zachmann, Calligra-Devel-list

Differential Revision: https://phabricator.kde.org/D128