Added an optional optimization to slider spin box
We are not always interested in constant updates through signal
when dragging the slider with the mouse or a tablet pen, so calling
setBlockUpdateSignalOnDrag(true) will inform the widget that
it has to internally update the slider value and ask inheriting classes
(in this case KisDoubleSliderSpinBox and KisSliderSpinBox)
to not send any valueChanged signal.

Since an update is sent anyway when releasing the left mouse button,
everything still works fine and for instance  now there's no slowdown
when dragging the brush size slider.

Currently that option is set to true to the Opacity, Flow and Size slider of KisPaintOpBox.

CCMAIL: kimageshop@kde.org