Fix assert when restoring selection during a document switch
The list of selected nodes is not cleared when the underlying dummies graph changes
so when the layerbox model is reset it tries to restore non existing nodes.

The node clearing has been moved inside KisLayerBox::setCanvas function, so we can remove
the two clearing in unsetCanvas and notifyImageDeleted since they would be unnecessary.

CCBUG: 349846