Change exactBounds() to handle non-transparent default pixel properly
Now the semantics of exactBounds() has changed. In the case when
the default pixel is not fully transparent, now we do not return the
extent of the device. Instead, we return either the image bounds value,
or (if the device is bigger than the image) device's nonDefaultPixelArea().
Technically it looks like a union
(defaultBounds()->bounds() | nonDefaultPixelArea()).

This fixes a bug when exactBounds() were gradually growing during the
merge operations in:

BUG:344012