Implement smooth scaling of the level of detail planes
This improves quality of the scaled planes, especially
on lower zoom levels.

Now we blend all the pixels involving in the destination pixel. The speed
is not ideal though. Generation of a plane for 3kx2k image take about 250 ms.

Probably, we will need to add an option for disabling it in the future.