Import a copy of KShortcutsDialog sources for disabling its scheme editor
Summary:
For the kdelibs4-based version of KisPart::configureShortcuts()
a hack was used to hide the shortcut schemes widget in the
normal shortcut editor KShortcutsDialog from kdelibs, by setting the
bottom buttons oneself. This does not work anymore (as the buttons are
no longer exposed), so for now running with a plain copy of the sources
of KShortcutsDialog, where the schemes editor is disabled directly.
Not nice, but then soon custom Krita-specific shortcut handling is
planned anyway, right?

Reviewers: dkazakov, langkamp, rempt

Differential Revision: https://phabricator.kde.org/D156