Remove KisOpenGLCanvas2::paintEvent
This is no longer needed after the port to QOpenGLWidget (see Qt
dox, section "Relation to QGLWidget":

"While the API is very similar, there is an important difference between
the two: QOpenGLWidget always renders offscreen, using framebuffer
objects. QGLWidget on the other hand uses a native window and surface. The
latter causes issues when using it in complex user interfaces since,
depending on the platform, such native child widgets may have various
limitations, regarding stacking orders for example. QOpenGLWidget avoids
this by not creating a separate native window.

Due to being backed by a framebuffer object, the behavior of QOpenGLWidget
is very similar to QOpenGLWindow with the update behavior set to
PartialUpdateBlit or PartialUpdateBlend. This means that the contents
are preserved between paintGL() calls so that incremental rendering is
possible. With QGLWidget (and naturally QOpenGLWindow with the default
update behavior) this is usually not the case because swapping the
buffers leaves the back buffer with undefined contents."