BUG:350650 Do not abort when saving jpeg2000
If you try to save to jpeg2000 (which you shouldn't...), the
extension encodes a particular type of jpeg2000. If we end up
in the jp2000 export code with the wrong extension, in a debug
build, we'd qFatal...

Since lots of people are using debug builds this is a bit rude.
Instead return -1 and _check_ the return code for validity.