BUG:337274 Filter the whole active node
Instead of limiting the filter action to the image. I chose to do
this instead of adding an option because the when I was doing that
I realized that it would just clutter the ui for very little
benefit.

But I do think that people should start using filter masks more
than directly and destructively applying filters.