Stamp and Clipboard brush fixes
Corrected KisGbrBrush copy constructor, the rhs private d should be passed
to the new Private constructor, not assigned, otherwise its fields won't be copied.

Stamp and Clipboard 'Color as mask' option becomes 'Create mask from color'
to avoid confusion with the 'Color as mask' option in the Predefine tab.

Fixed the preview for Stamp and Clipboard widgets when modifying the brush,
now it is correctly updated. Code should use brushTipImage() not image().