Convert colorants and whitepoints with the chromatic adaption matrix if there.
This is necessary because newer ICC profiles make the whitepoint d50 by
default and put a transform matrix from the actual white point to d50 into
the chromatic adaption tag. By inverting and converting we can actually
see our sRGB with D65 and CIERGB with E whitepoints. There's some rounding
issues, but this is good for now.