Remove misleading error message after saving fails
Unless we can determine whether the image is still processing something,
we shouldn't suggest it is. This error message happens every time a
filter fails without setting an error message, which is misleading in
99% of the cases.

Next: check all our filters to properly set an error message.

And then we need to rewrite the filter system to be not so incredibly
complicated. Probably better reserve that for 3.x though.

CCBUG:347850