Fix bundle and preset loading
On Windows the code to parse the bundle name and the relative path to the
filename inside the bundle fails because columns are used as a reference
tp split the string and the partition labels contains columns.
Since we are interested in the last ones anyway, we use lastIndexOf
instead of indexOf.