Fix bundle path and resource relative path parsing
This fixes what a5309e41f7d8c39d74526f07ad62a17c49df6591 commit did not fix.
On Linux colons (columns should be colons in the previous commit too)
are fine to have in filename while on Windows it is not.
The parsing code then has issues if colons are in the name, we then
search for the full string ".bundle:" which should be hopefully not
present inside a path or filename.