Accept and process tablet events by default
This is the central change defining D270.  It changes the main strategy
for processing events from removing events from the queue and relying on
the synthetic events, to a strategy of accepting as many events as
possible and sending them through the queue.

Most of the original processing logic could be reproduced inside the
new Event Eater found in kis_input_manager_p.cpp.