Move the karbon filter and tool plugins to calligra/plugins
Summary:
These are flake plugins that actually have very little ties with
karbon and are present and usable in other applications as well.
This fills a todo we decided in, I think 2007, in Berlin, in the
old KDAB office:

https://dot.kde.org/2007/10/30/second-koffice-sprint-berlin-focuses-release-polish

The gradient edit widget is now in kowidgets, the gradient helper
in flake, the plugins in calligra/plugins/karbonplugins.

Test Plan:

Reviewers:

Subscribers: