PictureShapeConfigWidget: add error handling to KIO::Job.
Mistyping a filename in the file widget would lead to the job failing silently
and the rest of the code trying to load an empty bytearray as the image data.

Better show a msgbox on error and delete the shape.

But this is last resort (e.g. remote URL and lost internet during transfer),
see next commit.

REVIEW: 125493