Sort the pages so the list is the same every time.
This fixes the problem:
When working on getting cstester ported I noticed that the pages are in a
random order all the time.  Looking at the code I see that the pages are stored
in a QHash and it seems that at least in Qt5 the order is not always the same
in every run.
In the header it is says that the pages are orderd.