Add waiting feedback when the user tries to save a busy image
When an image is calculating some action, saving is impossible,
so before we were just cancelling  it. Now we added a special waiting dialog that
shows the user that the image is doing something and he should wait.

There is also a button for cancelling currently running image action.

BUG:350283,354946
Fixes T592

Patch backported from 3.0.