Fix bug 359288
The problem is that the zIndex of the svg shape is never set and
therefore it has the same zIndex 0 as the already existing shape. The
problem can be seen always if you create a document with 3 vector
shapes where the text shape is the most top one.
This commit loads the zIndex also for svg shapes.

BUG: 359288