Fix loading of filllayers
It seems that the saving of the fill layers writes out the files in the
current locale thanks to QDomElement.
So for me the color of the fill layer is saved as:

<!DOCTYPE params>
<params version="1">
 <param type="color" name="color"><![CDATA[<!DOCTYPE color>
  <color>
    <RGB space="sRGB-elle-V2-g10.icc" g="0,5270923972129822"
    r="0,5271534323692322" b="6,103608757257462e-05"/>
  </color>
]]></param>
</params>

The behaviour of QString has been changed in this regard. In 4.8
it first tried the current locale and then C locale. In 5.x only
the C locale is used to convert to double.
The attached patch fixes the problem by using available functions
to convert the doubles.