Text layout: Fix crash due to use of invalid FrameIterator
Summary:
To keep app responsive layouting is chopped up in parts and
an iterator is used to remember where to continue.
In some (unknown) situations the QTextFrame the iterator refers to is deleted.
This patch implements an isValid() method that is checked before the iterator is used.

The crash is easily triggered by adding a chart to words or sheets and
eg show grid or hide axis label.

Test Plan: Tested by creating complex document in words, checking that layout is sane and complete.

Reviewers: boemann

Reviewed By: boemann

Tags: #kexi, #calligra:_3.0

Differential Revision: https://phabricator.kde.org/D3468