Add release script
This script creates a config.ini file with a list of all po file names
so a tarball can be produces using the standard create_tarball_kf5.rb script.

This is a temporary solution as the po file name extraction is very brittle,
and it taxes the svn server a lot.