Fix crash in chart, add a deleable flag to KoShape
Summary:
Problem is that ChartShape does not handle deleted components,
but it shall be possible to manipulate these components.

This patch adds a deletable flag to KoShape and checks this flag
in the default tool before trying to delete a shape.
ChartShape then sets this falg for all its components.

Currently isGeometryProtected() is used to block deletion of a shape.
This behavior is not changed atm.
Imo this should be reviewed so that geometry can be protected also for deletable shapes.

Reviewers: boemann, mecir, anthonyfieroni

Tags: #calligra:_3.0

Differential Revision: https://phabricator.kde.org/D4885