cmake: switch to the FindPoppler module from ECM
Bump the ECM requirement to 5.19, which provides a better FindPoppler
than the local copy (which was itself copied from okular).
As result, we can search for the Core and Qt5 components specifically,
and use the Poppler::Core and Poppler::Qt5 target where needed, with
no more manual include directories.

The only behaviour change is that pdf2svg does not link to poppler-qt5
anymore, as it only needs the private libpoppler.