Paint crop rectangle and its handles with 1px wide outline.
Due to the painter scaling the pen width (1 by default) was scaled too, which caused the outlines to cover the whole image.

This patch sets the pen width to 0 which always results in a 1px wide line, unfortunately also on HiDPI-Displays.
Using a pen width of 1 and changing everything to work with absolute coordinates to fix that seemed a little cumbersome, so I added a dark overlay for the cropped region to improve visibility instead.

Patch by Sebastian Pettke.

BUG: 388930
FIXED-IN: 3.1.0