Fix tool is not shown first time it is activated
First time the tool is activated, it is not shown
because activetool must be set before optionwidgets are set.

This patch works around this by setting option widgets also
when shape(s) is selected as this is the first time
activetool is set. This may cause an extra layout in certain cases,
but imho it is a small price to pay.