Chart: Major rework of (mainly) layout and ui
The ui is split up into separate widgets to look and work
more like other tool dockers in calligra.
The intention is to have only basic options in the docker to
keep it at a reasonable size. Other options will go into separate dialogs (todo).

Layout info is added to the ui (and odf) and is used in the layout code.

There are tons to do still (more options, bugs) but it should be possible
to do this in smaller steps from now on.