Add missing dependencies for PART_STAGE
KPrNotes is using the text shape and the picture shape in its constructor method
(see line 67 and line 78). As a consequence, in a presentation, when creating a
new page in KPrDocument.cpp#171, the page creation itself is creating a KPrNotes
(see KPrPage.cpp#56) and finally if one of the two shape plugins is not compiled,
the Q_ASSERT() will catch the non existing shape, at best, or if compiled with
QT_NO_DEBUG, the program will crash.

Also update two function calls using deprecated QMouseEvent::posF() in
plugins/pictureshape.

Reviewed By: leinir

Differential Revision: https://phabricator.kde.org/D15679