Make RecentFileManager less potentially crashy
This is a funky little situation where some of the assumptions
are just a little off. In short: There are going to be multiple
files with the same file name in different locations on the
filesystem. This will result in two lists ending up out of sync,
if we simply clear multiple instances of the same string. Solution:
switch to a less assumptioney and more strict container.