Extract images from excel files which omit the FtCF entry.
Summary: The specification says it and the flags entry must exist if the object type is picture but appears to not be true of some files. Since there's no essential information there add a basic check for the field id and if it not what is expected skip parsing of that field without invalidating the entire object.

Reviewers: pvuorela

Reviewed By: pvuorela

Subscribers: Calligra-Devel-list

Tags: #calligra:_3.0

Differential Revision: https://phabricator.kde.org/D25033