Read xlsx drawing dimensions from the xdr:ext element if available.
Summary: Read the size of a single cell anchored drawing from the xdr:ext element. The only real difference from the a:ext element is the namespace but that means the existing read_ext function fails reading xdr:ext and a new one can't be added without clashing so the logic is included inline instead of utilizing the usual macros.

Reviewers: pvuorela

Reviewed By: pvuorela

Subscribers: Calligra-Devel-list

Tags: #calligra:_3.0

Differential Revision: https://phabricator.kde.org/D25032