Some virtual funcions cannot be marked override
Seems gcc and clang gets confused if a pure virtual in class A
that is privately inherited in a sub-class B and overriden by a funtion marked override,
and B is subclassed by class C.
Result is that in some situations the wrong function is called.

Disclamer: There might be further conditions that needs to be fullfilled
that I have not been able to deduce.

Testing:
Unit test TestShapeAnimations passes.
Animations still works.