BKO#239200: missing backgrounds in docx files
There was a typo in the parsing of the settings from docx. This typo disabled loading the background in most cases.
Another, harder to solve issue, is the support of pattern backgrounds. Since there is no direct translation from
MS-OOXML to ODF here, a lot more code will be needed.