Fix crash on exit
A crash may occure because the canvas accesses the canvas controller after
it has been deleted.

This can appen because an inputmethod event is received after the controller is deleted.

Words always crashes, crash in stage/karbon can be reproduced as follows:
1) Add a shape to stage.
2) Close with save
3) Open document
4) Select shape
5) Close document -> crash

In general you need to have selected a shape at least once (deselecting doesn't matter).
The trick is words automatically selects a shape as pages are actually textshapes.
In stage/karbon you have to do it manually.