Move the document statistics to its own class
So far, the statistics computation was done directly
in KWDocumentStatistics. This made it impossible for
any other interface (QML for instance) to use these
statistics.

There is now a KWDocumentStatistics instance bound to
KWDocument. This instance will compute the statistics
if they are being required by another component. This
is known by looking at the listeners of the refreshed()
signal.