Results of testing compilation using Debian stretch
We got a report on IRC that building was broken (thanks 'jaquino'). This is my fault,
I fixed warning and did not realize I introduced a dependency on Qt 5.14 while we ask
only for 5.6 in CMakeLists.txt.
I had to do these changes to make Calligra build on Debian stretch, which is more or
less matching our minimal requirements (except it has Qt 5.7 instead of 5.6).
So as long as we build on stretch, we should not be breaking our requirements promise.