Work around QTBUG-92153
QTextDocument (at least for Qt ≤6.1) has a QSet<QTextCursor> to maintain all the cursors in a text
that should not move when content is inserted before their positions.
This structure is deeply flawed: when appending text to the document (typically when loading a
document for us), each cursor will be visited for a noop.
This is extremely expensive and is more than 20% of the loading time of the ODF spec on my
computer.
This patch works around this issue by loading the KoTextRanges by position first, and once
the document is fully loaded, then finalizing them by converting to QTextCursors.