Fix undefined behaviour in generateKey() and improve spread
The shift operation promotes the byte value to an int, however that's
only 32 bits and potentially -ve if char is signed so we end up
shifting more than the size of the type or shifting a -ve number. Both
of these are UB so add a cast to make sure that we operate on the
correct type.